Features state added to History

+++ src/main/scala/com/wavesplatform/features/FeatureProvider.scala
@@ -1,5 +1,5 @@
 package com.wavesplatform.features
 
 trait FeatureProvider {
-  def status(feature: Int): FeatureStatus
+  def status(feature: Short): FeatureStatus
 }
